refactor: move inline tests to engine/tests/ for state, clock, music_theory
