+++ engine/src/clock.rs
@@ -22,14 +22,14 @@ const BPM_MIN: u16 = 20;
 const BPM_MAX: u16 = 300;
 
 /// Initial seed for the clock-local Xorshift64 RNG (separate from `SequencerState::rng_seed`).
-const CLOCK_RNG_INIT: u64 = 0xA24B_AED4_963D_37C5;
+pub const CLOCK_RNG_INIT: u64 = 0xA24B_AED4_963D_37C5;
 
 // ── Clock-local tempo jitter types ──────────────────────────────────────────
 
 /// Clock-local tempo jitter state.
 ///
 /// Not stored in `SequencerState` — no lock needed.
-pub(crate) struct TempoRollState {
+pub struct TempoRollState {
     /// Step counter used for phase-based roll points (Beat, Seq).
     phase: u64,
     /// Current signed direction for PingPong (+1 or -1).
@@ -45,7 +45,7 @@ impl Default for TempoRollState {
 }
 
 /// Snapshot of tempo randomness params read from `SequencerState` under a read lock.
-pub(crate) struct TempoRandSnapshot {
+pub struct TempoRandSnapshot {
     /// Probability (0–100) that the tempo randomness roll fires.
     pub tempo_rand: u8,
     /// When the tempo randomness roll fires.
@@ -95,7 +95,7 @@ fn prob_hit(seed: &mut u64, chance: u8) -> bool {
 /// `step_count` is the total steps elapsed since clock start (for Beat/Seq roll points).
 ///
 /// Returns the effective BPM clamped to 20–300.
-pub(crate) fn compute_effective_bpm(
+pub fn compute_effective_bpm(
     base_bpm: u16,
     roll_state: &mut TempoRollState,
     params: &TempoRandSnapshot,
@@ -384,398 +384,6 @@ pub fn run_clock(state: Arc<RwLock<SequencerState>>, midi_tx: SyncSender<MidiEve
         // Advance next absolute time by one tick period.
         next_abs = add_nanos(next_abs, period);
         step_count = step_count.wrapping_add(1);
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::state::{MidiEvent, SequencerState, StepData, TempoRandType, TempoRollPoint};
-    use std::sync::mpsc;
-
-    /// Simulates the retrigger logic from `run_clock` on a single tick.
-    ///
-    /// Returns the events that would be emitted for the given `maybe_event`,
-    /// given the current `last_note` state.  Updates `last_note` in place,
-    /// mirroring the production code.
-    fn simulate_tick(
-        maybe_event: Option<MidiEvent>,
-        last_note: &mut Option<(u8, u8)>,
-        period: u64,
-    ) -> Vec<MidiEvent> {
-        let mut events = Vec::new();
-        if let Some(MidiEvent::NoteOn { channel, note, velocity, .. }) = maybe_event {
-            if *last_note == Some((channel, note)) {
-                events.push(MidiEvent::NoteOff { channel, note });
-            }
-            events.push(MidiEvent::NoteOn { channel, note, velocity, duration_nanos: period });
-            *last_note = Some((channel, note));
-        }
-        events
-    }
-
-    // ── BUG-018: repeated-note retrigger ────────────────────────────────────
-
-    /// Two consecutive steps with the same note must each produce a NoteOn,
-    /// with a NoteOff inserted between them.
-    #[test]
-    fn test_retrigger_same_note_inserts_note_off() {
-        let mut state = SequencerState::default();
-        // Steps 0 and 1 both enabled with note 60, channel 0.
-        state.steps[0] = StepData { enabled: true, midi_note: 60, velocity: 100 };
-        state.steps[1] = StepData { enabled: true, midi_note: 60, velocity: 100 };
-        state.playing = true;
-        // playhead starts at 0; first tick advances to step 0.
-        // (tick() increments before reading: 0+1 wraps at 16 → step 1, no — let's
-        //  check: default playhead=0, tick advances next=1 when loop_active=false
-        //  and next<16, so sets playhead=1. We want steps 0,1 both enabled.)
-        // Re-initialise so playhead is at 15 (wraps to 0 on first tick).
-        state.playhead = 15;
-
-        let period: u64 = 500_000;
-        let mut last_note: Option<(u8, u8)> = None;
-
-        let e1 = simulate_tick(state.tick(), &mut last_note, period);
-        assert_eq!(e1, vec![MidiEvent::NoteOn { channel: 0, note: 60, velocity: 100, duration_nanos: period }],
-            "first tick: NoteOn only");
-        assert_eq!(last_note, Some((0, 60)));
-
-        let e2 = simulate_tick(state.tick(), &mut last_note, period);
-        assert_eq!(e2, vec![
-            MidiEvent::NoteOff { channel: 0, note: 60 },
-            MidiEvent::NoteOn  { channel: 0, note: 60, velocity: 100, duration_nanos: period },
-        ], "second tick: NoteOff then NoteOn for retrigger");
-    }
-
-    /// A different note on the second step must NOT produce a NoteOff first.
-    #[test]
-    fn test_no_retrigger_for_different_note() {
-        let mut state = SequencerState::default();
-        state.steps[0] = StepData { enabled: true, midi_note: 60, velocity: 100 };
-        state.steps[1] = StepData { enabled: true, midi_note: 62, velocity: 100 };
-        state.playing = true;
-        state.playhead = 15;
-
-        let period: u64 = 500_000;
-        let mut last_note: Option<(u8, u8)> = None;
-
-        let e1 = simulate_tick(state.tick(), &mut last_note, period);
-        assert_eq!(e1, vec![MidiEvent::NoteOn { channel: 0, note: 60, velocity: 100, duration_nanos: period }]);
-
-        let e2 = simulate_tick(state.tick(), &mut last_note, period);
-        assert_eq!(e2, vec![MidiEvent::NoteOn { channel: 0, note: 62, velocity: 100, duration_nanos: period }],
-            "different note: no NoteOff inserted");
-    }
-
-    /// A disabled step must NOT update last_note, so no phantom NoteOff fires.
-    #[test]
-    fn test_disabled_step_does_not_update_last_note() {
-        let mut state = SequencerState::default();
-        state.steps[0] = StepData { enabled: true,  midi_note: 60, velocity: 100 };
-        state.steps[1] = StepData { enabled: false, midi_note: 60, velocity: 100 };
-        state.steps[2] = StepData { enabled: true,  midi_note: 60, velocity: 100 };
-        state.playing = true;
-        state.playhead = 15;
-
-        let period: u64 = 500_000;
-        let mut last_note: Option<(u8, u8)> = None;
-
-        let _e1 = simulate_tick(state.tick(), &mut last_note, period); // step 0 → NoteOn
-        let e2 = simulate_tick(state.tick(), &mut last_note, period); // step 1 → disabled, None
-        assert!(e2.is_empty(), "disabled step emits no events");
-        assert_eq!(last_note, Some((0, 60)), "last_note unchanged after disabled step");
-
-        // Step 2 has the same note: because last_note is still set, a NoteOff fires.
-        // This is correct behaviour — the note from step 0 is still logically held.
-        let e3 = simulate_tick(state.tick(), &mut last_note, period); // step 2 → retrigger
-        assert_eq!(e3, vec![
-            MidiEvent::NoteOff { channel: 0, note: 60 },
-            MidiEvent::NoteOn  { channel: 0, note: 60, velocity: 100, duration_nanos: period },
-        ], "retrigger fires after disabled gap");
-    }
-
-    // ── run_clock integration: channel-based smoke test ─────────────────────
-
-    /// Two consecutive same-note steps both produce NoteOn events on the MIDI
-    /// channel, with a NoteOff sandwiched between them.
-    ///
-    /// This test drives `run_clock` in a real thread (sleep_until is a no-op on
-    /// non-Linux so the loop spins fast). We collect exactly 3 events and then
-    /// drop the receiver so the clock thread exits cleanly.
-    #[test]
-    fn test_run_clock_retrigger_via_channel() {
-        use std::sync::{Arc, RwLock};
-
-        let mut state = SequencerState::default();
-        // Two enabled steps with the same note; rest disabled.
-        state.steps[0] = StepData { enabled: true, midi_note: 60, velocity: 100 };
-        state.steps[1] = StepData { enabled: true, midi_note: 60, velocity: 100 };
-        state.playing = true;
-        state.playhead = 15; // wraps to 0 on first tick
-        // Very fast tempo and small step so the thread doesn't wait long.
-        state.tempo_bpm = 240;
-        state.step_size = StepSize::ThirtySecond;
-
-        let shared = Arc::new(RwLock::new(state));
-        // Capacity 3: thread will block after 3 events; dropping rx causes the
-        // next send to error, exiting the loop.
-        let (tx, rx) = mpsc::sync_channel::<MidiEvent>(3);
-
-        let shared_clone = Arc::clone(&shared);
-        let handle = std::thread::spawn(move || run_clock(shared_clone, tx));
-
-        // Collect the first 3 events.
-        let ev1 = rx.recv().expect("event 1");
-        let ev2 = rx.recv().expect("event 2");
-        let ev3 = rx.recv().expect("event 3");
-        // Drop receiver — causes the clock thread to exit on its next send.
-        drop(rx);
-        handle.join().ok();
-
-        assert!(matches!(ev1, MidiEvent::NoteOn { note: 60, .. }), "ev1 NoteOn");
-        assert!(matches!(ev2, MidiEvent::NoteOff { note: 60, .. }), "ev2 NoteOff retrigger");
-        assert!(matches!(ev3, MidiEvent::NoteOn { note: 60, .. }), "ev3 NoteOn again");
-    }
-
-    // ── compute_effective_bpm: tempo_rand = 0 always returns base ───────────
-
-    #[test]
-    fn test_compute_effective_bpm_tempo_rand_zero_returns_base() {
-        let mut roll_state = TempoRollState::default();
-        let mut rng = CLOCK_RNG_INIT;
-        let params = TempoRandSnapshot {
-            tempo_rand: 0,
-            roll_point: TempoRollPoint::Step,
-            variance_max: 20,
-            rand_type: TempoRandType::Random,
-        };
-        for step in 0..1000u64 {
-            let effective = compute_effective_bpm(120, &mut roll_state, &params, &mut rng, step);
-            assert_eq!(effective, 120, "tempo_rand=0 must always return base BPM");
-        }
-    }
-
-    // ── compute_effective_bpm: roll_point = Off always returns base ──────────
-
-    #[test]
-    fn test_compute_effective_bpm_off_returns_base() {
-        let mut roll_state = TempoRollState::default();
-        let mut rng = CLOCK_RNG_INIT;
-        let params = TempoRandSnapshot {
-            tempo_rand: 100,
-            roll_point: TempoRollPoint::Off,
-            variance_max: 20,
-            rand_type: TempoRandType::Random,
-        };
-        for step in 0..1000u64 {
-            let effective = compute_effective_bpm(120, &mut roll_state, &params, &mut rng, step);
-            assert_eq!(effective, 120, "roll_point=Off must always return base BPM");
-        }
-    }
-
-    // ── compute_effective_bpm: Random stays within base ± variance_max ───────
-
-    #[test]
-    fn test_compute_effective_bpm_random_stays_in_bounds() {
-        let mut roll_state = TempoRollState::default();
-        let mut rng = CLOCK_RNG_INIT;
-        let base = 120u16;
-        let vm = 20u8;
-        let params = TempoRandSnapshot {
-            tempo_rand: 100,
-            roll_point: TempoRollPoint::Step,
-            variance_max: vm,
-            rand_type: TempoRandType::Random,
-        };
-        for step in 0..1000u64 {
-            let effective = compute_effective_bpm(base, &mut roll_state, &params, &mut rng, step);
-            assert!(
-                effective >= base - vm as u16 && effective <= base + vm as u16,
-                "Random BPM {effective} out of [{}, {}]",
-                base - vm as u16, base + vm as u16
-            );
-        }
-    }
-
-    // ── compute_effective_bpm: effective BPM clamped to 20–300 ──────────────
-
-    #[test]
-    fn test_compute_effective_bpm_clamped_to_range() {
-        let mut roll_state = TempoRollState::default();
-        let mut rng = CLOCK_RNG_INIT;
-        // Base near floor; large variance_max → clamp at BPM_MIN=20
-        let params = TempoRandSnapshot {
-            tempo_rand: 100,
-            roll_point: TempoRollPoint::Step,
-            variance_max: 99,
-            rand_type: TempoRandType::Random,
-        };
-        for step in 0..1000u64 {
-            let effective = compute_effective_bpm(25, &mut roll_state, &params, &mut rng, step);
-            assert!(effective >= 20, "BPM {effective} below minimum 20");
-            assert!(effective <= 300, "BPM {effective} above maximum 300");
-        }
-    }
-
-    // ── compute_effective_bpm: PingPong bounces monotonically ────────────────
-
-    #[test]
-    fn test_compute_effective_bpm_pingpong_bounces() {
-        let mut roll_state = TempoRollState::default();
-        let mut rng = CLOCK_RNG_INIT;
-        let base = 120u16;
-        let vm = 10u8;
-        let params = TempoRandSnapshot {
-            tempo_rand: 100,
-            roll_point: TempoRollPoint::Step,
-            variance_max: vm,
-            rand_type: TempoRandType::PingPong,
-        };
-
-        let mut prev = base as i32;
-        let mut direction_up = true;
-
-        for step in 0..200u64 {
-            let effective = compute_effective_bpm(base, &mut roll_state, &params, &mut rng, step) as i32;
-            let offset = effective - base as i32;
-            assert!(offset >= -(vm as i32) && offset <= vm as i32,
-                "PingPong offset {offset} out of bounds at step {step}");
-
-            // Verify monotonicity within each half-sweep.
-            if direction_up {
-                if effective < prev {
-                    // Direction reversal — we hit the top.
-                    direction_up = false;
-                }
-            } else if effective > prev {
-                // Direction reversal — we hit the bottom.
-                direction_up = true;
-            }
-            prev = effective;
-        }
-    }
-
-    // ── compute_effective_bpm: Breathe forms triangle wave within bounds ─────
-
-    #[test]
-    fn test_compute_effective_bpm_breathe_within_bounds() {
-        let mut roll_state = TempoRollState::default();
-        let mut rng = CLOCK_RNG_INIT;
-        let base = 120u16;
-        let vm = 10u8;
-        let params = TempoRandSnapshot {
-            tempo_rand: 100,
-            roll_point: TempoRollPoint::Step,
-            variance_max: vm,
-            rand_type: TempoRandType::Breathe,
-        };
-        for step in 0..1000u64 {
-            let effective = compute_effective_bpm(base, &mut roll_state, &params, &mut rng, step);
-            let offset = effective as i32 - base as i32;
-            assert!(
-                offset >= -(vm as i32) && offset <= vm as i32,
-                "Breathe offset {offset} out of bounds at step {step}"
-            );
-        }
-    }
-
-    // ── tempo_bpm never mutated by clock thread ───────────────────────────────
-
-    #[test]
-    fn test_tempo_bpm_never_mutated_by_run_clock() {
-        use std::sync::{Arc, RwLock};
-
-        let initial_bpm = 120u16;
-        let mut state = SequencerState::default();
-        state.tempo_bpm = initial_bpm;
-        state.tempo_rand = 100;
-        state.tempo_roll_point = TempoRollPoint::Step;
-        state.tempo_variance_max = 20;
-        state.tempo_rand_type = TempoRandType::Random;
-        state.playing = true;
-        state.steps[0] = StepData { enabled: true, midi_note: 60, velocity: 100 };
-        state.tempo_bpm = 240;
-        state.step_size = StepSize::ThirtySecond;
-
-        let shared = Arc::new(RwLock::new(state));
-        let (tx, rx) = mpsc::sync_channel::<MidiEvent>(32);
-
-        let shared_clone = Arc::clone(&shared);
-        let handle = std::thread::spawn(move || run_clock(shared_clone, tx));
-
-        // Collect 10 events then drop receiver.
-        for _ in 0..10 {
-            let _ = rx.recv();
-        }
-        drop(rx);
-        handle.join().ok();
-
-        let final_bpm = shared.read().expect("read").tempo_bpm;
-        assert_eq!(final_bpm, 240, "tempo_bpm must not be mutated by the clock thread");
-    }
-
-    // ── Seq roll point fires every 16 steps only ─────────────────────────────
-
-    #[test]
-    fn test_compute_effective_bpm_seq_fires_every_16_steps() {
-        // With Seq roll point and tempo_rand=100, jitter should only change
-        // at multiples of 16. Between those boundaries the offset stays constant.
-        let mut roll_state = TempoRollState::default();
-        let mut rng = CLOCK_RNG_INIT;
-        let base = 120u16;
-        let params = TempoRandSnapshot {
-            tempo_rand: 100,
-            roll_point: TempoRollPoint::Seq,
-            variance_max: 10,
-            rand_type: TempoRandType::Up,
-        };
-        let mut last_bpm = compute_effective_bpm(base, &mut roll_state, &params, &mut rng, 0);
-
-        for step in 1..80u64 {
-            let effective = compute_effective_bpm(base, &mut roll_state, &params, &mut rng, step);
-            if effective != last_bpm {
-                // A change must only happen at a multiple of 16.
-                assert_eq!(step % 16, 0,
-                    "Seq roll changed at step {step}, not a multiple of 16");
-                last_bpm = effective;
-            }
-        }
-    }
-
-    // ── Beat roll point fires every 4 steps only ─────────────────────────────
-
-    #[test]
-    fn test_compute_effective_bpm_beat_fires_every_4_steps() {
-        // With Beat roll point and tempo_rand=100, jitter should only change
-        // at multiples of 4. Between those boundaries the offset stays constant.
-        let mut roll_state = TempoRollState::default();
-        let mut rng = CLOCK_RNG_INIT;
-        let base = 120u16;
-        let params = TempoRandSnapshot {
-            tempo_rand: 100,
-            roll_point: TempoRollPoint::Beat,
-            variance_max: 10,
-            rand_type: TempoRandType::Up,
-        };
-        let mut last_update_step: Option<u64> = None;
-        let mut last_bpm = compute_effective_bpm(base, &mut roll_state, &params, &mut rng, 0);
-
-        for step in 1..40u64 {
-            let effective = compute_effective_bpm(base, &mut roll_state, &params, &mut rng, step);
-            if effective != last_bpm {
-                // A change must only happen at a multiple of 4.
-                assert_eq!(step % 4, 0,
-                    "Beat roll changed at step {step}, not a multiple of 4");
-                last_update_step = Some(step);
-                last_bpm = effective;
-            } else {
-                // Within a beat window, BPM must not change.
-                if let Some(last) = last_update_step {
-                    assert!(step - last < 4,
-                        "BPM unchanged for {} steps (expected change at beat boundary)", step - last);
-                }
-            }
-        }
     }
 }
