+++ engine/tests/music_theory.rs
@@ -1,4 +1,4 @@
-use engine::music_theory::{note_name, notes_in_key, next_note, SCALE_INTERVALS, Key, Mode};
+use engine::music_theory::{note_name, notes_in_key, next_note, snap_to_key, SCALE_INTERVALS, Key, Mode};
 
 #[test]
 fn all_modes_intervals_sum_to_12() {
@@ -303,4 +303,55 @@ fn next_note_off_key_in_g_major_up() {
     // F4 (65) is not in G Major. Just verify result is within MIDI range.
     let result = next_note(65, Key::G, Mode::Major, 1);
     assert!(result <= 127, "next_note result must be within MIDI range");
+}
+
+// --- snap_to_key ---
+
+#[test]
+fn snap_in_key_note_unchanged() {
+    // C major scale: C4=60, D4=62, E4=64, F4=65, G4=67, A4=69, B4=71
+    assert_eq!(snap_to_key(60, Key::C, Mode::Major), 60);
+    assert_eq!(snap_to_key(62, Key::C, Mode::Major), 62);
+    assert_eq!(snap_to_key(71, Key::C, Mode::Major), 71);
+}
+
+#[test]
+fn snap_out_of_key_rounds_to_nearest() {
+    // C# (61) is equidistant from C(60) and D(62) — tie: lower wins → C(60)
+    assert_eq!(snap_to_key(61, Key::C, Mode::Major), 60);
+    // Bb (70) in C major: nearest are A(69) dist=1, B(71) dist=1 — tie: lower wins → A(69)
+    assert_eq!(snap_to_key(70, Key::C, Mode::Major), 69);
+}
+
+#[test]
+fn snap_tie_picks_lower_note() {
+    // F# (66) in C major: F=65 (dist 1), G=67 (dist 1) — lower wins → F(65)
+    assert_eq!(snap_to_key(66, Key::C, Mode::Major), 65);
+}
+
+#[test]
+fn snap_across_octaves() {
+    // C5 = 72 is in C major
+    assert_eq!(snap_to_key(72, Key::C, Mode::Major), 72);
+    // C#5 = 73: C5(72) dist=1, D5(74) dist=1 — tie: lower wins → C5(72)
+    assert_eq!(snap_to_key(73, Key::C, Mode::Major), 72);
+}
+
+#[test]
+fn snap_midi_boundaries() {
+    let _ = snap_to_key(0, Key::C, Mode::Major);
+    let _ = snap_to_key(127, Key::C, Mode::Major);
+}
+
+#[test]
+fn snap_non_c_key() {
+    // Ab/G# (68) in G major: G=67 (dist 1), A=69 (dist 1) — tie: lower wins → G(67)
+    assert_eq!(snap_to_key(68, Key::G, Mode::Major), 67);
+}
+
+#[test]
+fn snap_natural_minor() {
+    // A natural minor: A=69, B=71, C=72, D=74, E=76, F=77, G=79
+    // Bb (70) in A natural minor: A=69 (dist 1), B=71 (dist 1) — tie: lower wins → A(69)
+    assert_eq!(snap_to_key(70, Key::A, Mode::NaturalMinor), 69);
 }
